Reduced some duplication in `GradleSamplesRunner` and bust out `GradleRunnerCommandExecutor` so that it can be reused in other contexts.
