+++ sample-check/src/main/java/org/gradle/samples/test/runner/GradleSamplesRunner.java
@@ -15,32 +15,25 @@
  */
 package org.gradle.samples.test.runner;
 
-import org.apache.commons.io.IOUtils;
 import org.gradle.api.JavaVersion;
 import org.gradle.samples.executor.CliCommandExecutor;
-import org.gradle.samples.executor.CommandExecutionResult;
 import org.gradle.samples.executor.CommandExecutor;
 import org.gradle.samples.executor.ExecutionMetadata;
+import org.gradle.samples.executor.GradleRunnerCommandExecutor;
 import org.gradle.samples.model.Command;
 import org.gradle.samples.model.Sample;
-import org.gradle.testkit.runner.BuildResult;
-import org.gradle.testkit.runner.GradleRunner;
 import org.junit.Rule;
 import org.junit.rules.TemporaryFolder;
 import org.junit.runners.model.InitializationError;
 
+import javax.annotation.Nullable;
 import java.io.File;
-import java.io.OutputStream;
-import java.io.OutputStreamWriter;
-import java.io.Writer;
-import java.util.ArrayList;
-import java.util.List;
 
 /**
  * A custom implementation of {@link SamplesRunner} that uses the Gradle Tooling API to execute sample builds.
  */
 public class GradleSamplesRunner extends SamplesRunner {
-    public static final String GRADLE_EXECUTABLE = "gradle";
+    private static final String GRADLE_EXECUTABLE = "gradle";
     @Rule
     public TemporaryFolder tempGradleUserHomeDir = new TemporaryFolder();
     private File customGradleInstallation = null;
@@ -61,37 +54,26 @@ public class GradleSamplesRunner extends SamplesRunner {
     }
 
     @Override
-    public CommandExecutionResult execute(File tempSampleOutputDir, File workingDir, Command command) {
+    protected CommandExecutor selectExecutor(ExecutionMetadata executionMetadata, File workingDir, Command command) {
         boolean expectFailure = command.isExpectFailure();
-        ExecutionMetadata executionMetadata = getExecutionMetadata(tempSampleOutputDir);
-        return new GradleRunnerCommandExecutor(workingDir, customGradleInstallation, expectFailure).execute(command, executionMetadata);
+        if (command.getExecutable().equals(GRADLE_EXECUTABLE)) {
+            return new GradleRunnerCommandExecutor(workingDir, customGradleInstallation, expectFailure);
+        }
+        return new CliCommandExecutor(workingDir);
     }
 
+    @Nullable
     @Override
-    protected File getSamplesRootDir() {
-        final String gradleHomeDir = getCustomGradleInstallationFromSystemProperty();
-        SamplesRoot samplesRoot = getTestClass().getAnnotation(SamplesRoot.class);
-        File samplesRootDir;
-        try {
-            if (samplesRoot != null) {
-                samplesRootDir = new File(samplesRoot.value());
-            } else if (System.getProperty("integTest.samplesdir") != null) {
-                String samplesRootProperty = System.getProperty("integTest.samplesdir", gradleHomeDir + "/samples");
-                samplesRootDir = new File(samplesRootProperty);
-            } else if (customGradleInstallation != null) {
-                samplesRootDir = new File(customGradleInstallation, "samples");
-            } else {
-                throw new InitializationError("Samples root directory is not declared. Please annotate your test class with @SamplesRoot(\"path/to/samples\")");
-            }
-
-            if (!samplesRootDir.exists()) {
-                throw new InitializationError("Samples root directory " + samplesRootDir.getAbsolutePath() + " does not exist");
-            }
-        } catch (InitializationError e) {
-            throw new RuntimeException("Could not initialize GradleSamplesRunner", e);
+    protected File getImplicitSamplesRootDir() {
+        String gradleHomeDir = getCustomGradleInstallationFromSystemProperty();
+        if (System.getProperty("integTest.samplesdir") != null) {
+            String samplesRootProperty = System.getProperty("integTest.samplesdir", gradleHomeDir + "/samples");
+            return new File(samplesRootProperty);
+        } else if (customGradleInstallation != null) {
+            return new File(customGradleInstallation, "samples");
+        } else {
+            return null;
         }
-
-        return samplesRootDir;
     }
 
     private String getCustomGradleInstallationFromSystemProperty() {
@@ -109,51 +91,4 @@ public class GradleSamplesRunner extends SamplesRunner {
         return gradleHomeDirProperty;
     }
 
-    private static class GradleRunnerCommandExecutor extends CommandExecutor {
-        private final File workingDir;
-        private final File customGradleInstallation;
-        private final boolean expectFailure;
-
-        private GradleRunnerCommandExecutor(File workingDir, File customGradleInstallation, boolean expectFailure) {
-            this.workingDir = workingDir;
-            this.customGradleInstallation = customGradleInstallation;
-            this.expectFailure = expectFailure;
-        }
-
-        @Override
-        protected int run(String executable, List<String> args, List<String> flags, OutputStream output) {
-            if (!executable.equals(GRADLE_EXECUTABLE)) {
-                return new CliCommandExecutor(workingDir).run(executable, args, flags, output);
-            }
-
-            List<String> allArguments = new ArrayList<>(args);
-            allArguments.addAll(flags);
-
-            GradleRunner gradleRunner = GradleRunner.create()
-                    .withProjectDir(workingDir)
-                    .withArguments(allArguments)
-                    .forwardOutput();
-
-            if (customGradleInstallation != null) {
-                gradleRunner.withGradleInstallation(customGradleInstallation);
-            }
-
-            Writer mergedOutput = new OutputStreamWriter(output);
-            try {
-                BuildResult buildResult;
-                if (expectFailure) {
-                    buildResult = gradleRunner.buildAndFail();
-                } else {
-                    buildResult = gradleRunner.build();
-                }
-                mergedOutput.write(buildResult.getOutput());
-                mergedOutput.close();
-                return expectFailure ? 1 : 0;
-            } catch (Exception e) {
-                throw new RuntimeException("Could not execute " + executable, e);
-            } finally {
-                IOUtils.closeQuietly(mergedOutput);
-            }
-        }
-    }
 }
